- Moved kernels and weights to their own files. - Modified PersistenceImager().__repr__ to create a valid constructor. - Removed PersistenceImager().dict_print and replaced with prettyprint. - Changed default kernel function name from bvncdf to gaussian to make more understandable. - Added a new test to ensure guassian parameter sigma can be either a numpy array or a list of lists. - Renamed some kernel functions by removing leading underscore to ensure proper loading. - Updated block comments in PersistenceImager() to reflect updated __repr__ method.
