+++ persim/images.py
@@ -5,15 +5,17 @@ import collections
 from joblib import Parallel, delayed
 
 import copy
+
 import numpy as np
 from scipy.stats import multivariate_normal as mvn
 from scipy.stats import norm
 import scipy.spatial as spatial
-from scipy.special import erfc
 import matplotlib.pyplot as plt
 from matplotlib.collections import LineCollection
+from persim.images_weights import *
+from persim.images_kernels import *
 
-from sklearn.base import TransformerMixin
+from sklearn.base import TransformerMixin, BaseEstimator
 
 __all__ = ["PersImage", "PersistenceImager"]
 
@@ -246,40 +248,25 @@ class PersistenceImager(TransformerMixin):
     >>> pimgr = PersistenceImager(pixel_size=0.2, birth_range=(0,1))
     ```
     
-    Printing a PersistenceImager() object will print its hyperparameters (defaults in this case):
+    Printing a PersistenceImager() object will print its hyperparameters:
     ```
     >>> print(pimgr)
 
-    PersistenceImager object:
-      pixel size: 0.2
-      resolution: (5, 5)
-      birth range: (0, 1)
-      persistence range: (0, 1)
-      weight: persistence
-      kernel: bvncdf
-      weight parameters: {n: 1.0}
-      kernel parameters: {sigma: [[ 1.  0.]
-                                  [ 0.  1.]]}
+        PersistenceImager(birth_range=(0.0, 1.0), pers_range=(0.0, 1.0), pixel_size=0.2, weight=persistence, weight_params={'n': 1.0}, kernel=gaussian, kernel_params={'sigma': [[1.0, 0.0], [0.0, 1.0]]})
+    ```
     
     PersistenceImager() attributes can be adjusted at or after instantiation. Updating attributes of a PersistenceImager() object will automatically update all other dependent attributes.
     ```
     >>> pimgr.pixel_size = 0.1
     >>> pimgr.birth_range = (0, 2)
     >>> print(pimgr)
-    
-    PersistenceImager object: 
-      pixel size: 0.1 
-      resolution: (20, 10)  <---
-      birth range: (0, 2) 
-      persistence range: (0, 1) 
-      weight: persistence 
-      kernel: bvncdf 
-      weight parameters: {n: 1.0} 
-      kernel parameters: {sigma: [[1. 0.]
-                                  [0. 1.]]}
+    >>> print(pimgr.resolution)
+        
+        PersistenceImager(birth_range=(0.0, 2.0), pers_range=(0.0, 1.0), pixel_size=0.1, weight=persistence, weight_params={'n': 1.0}, kernel=gaussian, kernel_params={'sigma': [[1.0, 0.0], [0.0, 1.0]]})
+        (20, 10)
     ```
     
-    The `fit()` method can be called on one or more (*,2) numpy arrays to automatically determine the miniumum birth and persistence ranges needed to capture all persistence pairs. The ranges and resolution are automatically adjusted to accomodate the specified pixel size.
+    The `fit()` method can be called on one or more (*,2) numpy arrays to automatically determine the miniumum birth and persistence ranges needed to capture all persistence pairs. The ranges and resolution are automatically adjusted to accomodate the specified pixel size. The option `skew=True` specifies that the diagram is currently in birth-death coordinates and must first be transformed to birth-persistence coordinates.
     
     ```
     >>> import numpy as np
@@ -288,19 +275,11 @@ class PersistenceImager(TransformerMixin):
                  np.array([[0.1, 0.2], [3.1, 3.3], [1.6, 2.9]]),
                  np.array([[0.2, 1.5], [0.4, 0.6], [0.2, 2.6]])]
     >>> pimgr.fit(pdgms, skew=True)
-    >>> pimgr
-    
-    PersistenceImager object: 
-      pixel size: 0.5 
-      resolution: (6, 5)                      <---
-      birth range: (0.1, 3.1)                 <---
-      persistence range: (-8.32667e-17, 2.5)  <---
-      weight: persistence 
-      kernel: bvncdf 
-      weight parameters: {n: 1.0} 
-      kernel parameters: {sigma: [[1. 0.]
-                                  [0. 1.]]}
-    
+    >>> print(pimgr)
+    >>> print(pimgr.resolution)
+        
+        PersistenceImager(birth_range=(0.1, 3.1), pers_range=(-8.326672684688674e-17, 2.5), pixel_size=0.5, weight=persistence, weight_params={'n': 1.0}, kernel=gaussian, kernel_params={'sigma': [[1.0, 0.0], [0.0, 1.0]]})
+        (6, 5)
     ```
     
     The `transform()` method can then be called on one or more (*,2) numpy arrays to generate persistence images from diagrams. The option `skew=True` specifies that the diagrams are currently in birth-death coordinates and must first be transformed to birth-persistence coordinates.
@@ -315,8 +294,6 @@ class PersistenceImager(TransformerMixin):
                [0.03625436, 0.05757744, 0.07242608, 0.07180125, 0.05593626],
                [0.02922239, 0.04712024, 0.05979033, 0.05956698, 0.04653357]])
     ```
-    
-    The option `skew=True` specifies that the diagram is currently in birth-death coordinates and must first be transformed to birth-persistence coordinates.
     """
     
     def __init__(self, birth_range=None, pers_range=None, pixel_size=None,
@@ -334,11 +311,11 @@ class PersistenceImager(TransformerMixin):
         if weight is None:
             weight = persistence
         if kernel is None:
-            kernel = bvncdf
+            kernel = gaussian
         if weight_params is None:
             weight_params = {'n': 1.0}
         if kernel_params is None:
-            kernel_params = {'sigma': np.array([[1.0, 0.0], [0.0, 1.0]])}
+            kernel_params = {'sigma': [[1.0, 0.0], [0.0, 1.0]]}
        
         self.weight = weight
         self.weight_params = weight_params
@@ -397,17 +374,12 @@ class PersistenceImager(TransformerMixin):
         self._height = int(np.ceil((self.pers_range[1] - self.pers_range[0]) / self.pixel_size)) * self._pixel_size
         self._resolution = (int(self.width / self.pixel_size), int(self.height / self.pixel_size))
         self._create_mesh()
-
+    
     def __repr__(self):
-        repr_str = 'PersistenceImager object: \n' +\
-                   '  pixel size: %g \n' % self.pixel_size +\
-                   '  resolution: (%d, %d) \n' % self.resolution +\
-                   '  birth range: (%g, %g) \n' % self.birth_range +\
-                   '  persistence range: (%g, %g) \n' % self.pers_range +\
-                   '  weight: %s \n' % self.weight.__name__ +\
-                   '  kernel: %s \n' % self.kernel.__name__ +\
-                   '  weight parameters: %s \n' % dict_print(self.weight_params) +\
-                   '  kernel parameters: %s' % dict_print(self.kernel_params)
+        import pprint as pp
+        
+        params = tuple([self.birth_range, self.pers_range, self.pixel_size, self.weight.__name__, pp.pformat(self.weight_params), self.kernel.__name__, pp.pformat(self.kernel_params)])
+        repr_str = 'PersistenceImager(birth_range=%s, pers_range=%s, pixel_size=%s, weight=%s, weight_params=%s, kernel=%s, kernel_params=%s)' % params
         return repr_str
 
     def _create_mesh(self):
@@ -514,7 +486,6 @@ class PersistenceImager(TransformerMixin):
 
         return pers_imgs
  
-
     def _ensure_iterable(self, pers_dgms):
         # if first entry of first entry is not iterable, then diagrams is singular and we need to make it a list of diagrams
         try:
@@ -609,18 +580,7 @@ class PersistenceImager(TransformerMixin):
         # optionally save figure
         if out_file:
             plt.savefig(out_file, bbox_inches='tight')
-
-def dict_print(dict_in):
-    # print dictionary contents in human-readable format
-    if dict_in is None:
-        str_out = 'None'
-    else:
-        str_out = []
-        for key, val in dict_in.items():
-            str_out.append('%s: %s' % (key, str(val)))
-        str_out = '{' + ', '.join(str_out) + '}'
-
-    return str_out
+            
 
 def _transform(pers_dgm, skew=True, resolution=None, weight=None, weight_params=None, kernel=None, kernel_params=None, _bpnts=None, _ppnts=None):
         """
@@ -644,7 +604,7 @@ def _transform(pers_dgm, skew=True, resolution=None, weight=None, weight_params=
         wts = weight(pers_dgm[:, 0], pers_dgm[:, 1], **weight_params)
 
         # handle the special case of a standard, isotropic Gaussian kernel
-        if kernel == bvncdf:
+        if kernel == gaussian:
             general_flag = False
             sigma = kernel_params['sigma']
 
@@ -652,11 +612,11 @@ def _transform(pers_dgm, skew=True, resolution=None, weight=None, weight_params=
             if isinstance(sigma, (int, float)):
                 sigma = np.array([[sigma, 0.0], [0.0, sigma]], dtype=np.float64)
 
-            if (sigma[0, 0] == sigma[1, 1] and sigma[0, 1] == 0.0):
-                sigma = np.sqrt(sigma[0, 0])
+            if (sigma[0][0] == sigma[1][1] and sigma[0][1] == 0.0):
+                sigma = np.sqrt(sigma[0][0])
                 for i in range(n):
-                    ncdf_b = _norm_cdf((_bpnts - pers_dgm[i, 0]) / sigma)
-                    ncdf_p = _norm_cdf((_ppnts - pers_dgm[i, 1]) / sigma)
+                    ncdf_b = norm_cdf((_bpnts - pers_dgm[i, 0]) / sigma)
+                    ncdf_p = norm_cdf((_ppnts - pers_dgm[i, 1]) / sigma)
                     curr_img = ncdf_p[None, :] * ncdf_b[:, None]
                     pers_img += wts[i]*(curr_img[1:, 1:] - curr_img[:-1, 1:] - curr_img[1:, :-1] + curr_img[:-1, :-1])
             else:
@@ -673,242 +633,3 @@ def _transform(pers_dgm, skew=True, resolution=None, weight=None, weight_params=
                 pers_img += wts[i]*(curr_img[1:, 1:] - curr_img[:-1, 1:] - curr_img[1:, :-1] + curr_img[:-1, :-1])
 
         return pers_img
-"""
-Kernel functions:
-
-A valid kernel is a python function of the form kernel(x, y, mu=(birth, persistence), **kwargs) defining a 
-cumulative distribution function such that kernel(x, y) = P(X <= x, Y <=y), where x and y are numpy arrays of equal length. 
-
-The required parameter mu defines the dependance of the kernel on the location of a persistence pair and is usually 
-taken to be the mean of the probability distribution function associated to kernel CDF.
-"""
-def uniform(x, y, mu=None, width=1, height=1):
-    w1 = np.maximum(x - (mu[0] - width/2), 0)
-    h1 = np.maximum(y - (mu[1] - height/2), 0)
-    
-    w = np.minimum(w1, width)
-    h = np.minimum(h1, height)
-
-    return w*h / (width*height)
-
-
-def bvncdf(birth, pers, mu=None, sigma=None):
-    """
-    Optimized bivariate normal cumulative distribution function for computing persistence images using a Gaussian kernel
-    :param birth: birth-coordinate(s) of pixel corners
-    :param pers: persistence-coordinate(s) of pixel corners
-    :param mu: (2,)-numpy array specifying x and y coordinates of distribution means (birth-persistence pairs)
-    :param sigma: (2,2)-numpy array specifying distribution covariance matrix or numeric if distribution is isotropic
-    :return: P(X <= birth, Y <= pers)
-    """
-    if mu is None:
-        mu = np.array([0.0, 0.0], dtype=np.float64)
-    if sigma is None:
-        sigma = np.array([[1.0, 0.0], [0.0, 1.0]], dtype=np.float64)
-
-    if sigma[0, 1] == 0.0:
-        return _sbvn_cdf(birth, pers,
-                         mu_x=mu[0], mu_y=mu[1], sigma_x=sigma[0, 0], sigma_y=sigma[1, 1])
-    else:
-        return _bvn_cdf(birth, pers,
-                        mu_x=mu[0], mu_y=mu[1], sigma_xx=sigma[0, 0], sigma_yy=sigma[1, 1], sigma_xy=sigma[0, 1])
-
-
-def _norm_cdf(x):
-    """
-    Univariate normal cumulative distribution function with mean 0.0 and standard deviation 1.0
-    :param x: value at which to evaluate the cdf (upper limit)
-    :return: P(X <= x), for X ~ N[0,1]
-    """
-    return erfc(-x / np.sqrt(2.0)) / 2.0
-
-
-def _sbvn_cdf(x, y, mu_x=0.0, mu_y=0.0, sigma_x=1.0, sigma_y=1.0):
-    """
-    Standard bivariate normal cumulative distribution function with specified mean and variances
-    :param x: x-coordinate(s) at which to evaluate the cdf (upper limit)
-    :param y: y-coordinate(s) at which to evaluate the cdf (upper limit)
-    :param mu_x: x-coordinate of mean of bivariate normal
-    :param mu_y: y-coordinate of mean of bivariate normal
-    :param sigma_x: variance in x
-    :param sigma_y: variance in y
-    :return: P(X <= x, Y <= y)
-    """
-    x = (x - mu_x) / np.sqrt(sigma_x)
-    y = (y - mu_y) / np.sqrt(sigma_y)
-    return _norm_cdf(x) * _norm_cdf(y)
-
-
-def _bvn_cdf(x, y, mu_x=0.0, mu_y=0.0, sigma_xx=1.0, sigma_yy=1.0, sigma_xy=0.0):
-    """
-    Bivariate normal cumulative distribution function with specified mean and covariance matrix based on the Matlab
-    implementations by Thomas H. Jørgensen (http://www.tjeconomics.com/code/) and Alan Genz
-    (http://www.math.wsu.edu/math/faculty/genz/software/matlab/bvnl.m) using the approach described by Drezner
-    and Wesolowsky (https://doi.org/10.1080/00949659008811236)
-    :param x: x-coordinate(s) at which to evaluate the cdf (upper limit)
-    :param y: y-coordinate(s) at which to evaluate the cdf (upper limit)
-    :param mu_x: x-coordinate of mean of bivariate normal
-    :param mu_y: y-coordinate of mean of bivariate normal
-    :param sigma_xx: variance in x
-    :param sigma_yy: variance in y
-    :param sigma_xy: covariance of x and y
-    :return: P(X <= x, Y <= y)
-    """
-    dh = -(x - mu_x) / np.sqrt(sigma_xx)
-    dk = -(y - mu_y) / np.sqrt(sigma_yy)
-
-    hk = np.multiply(dh, dk)
-    r = sigma_xy / np.sqrt(sigma_xx * sigma_yy)
-
-    lg, w, x = _gauss_legendre_quad(r)
-
-    dim1 = np.ones((len(dh),), dtype=np.float64)
-    dim2 = np.ones((lg,), dtype=np.float64)
-    bvn = np.zeros((len(dh),), dtype=np.float64)
-
-    if abs(r) < 0.925:
-        hs = (np.multiply(dh, dh) + np.multiply(dk, dk)) / 2.0
-        asr = np.arcsin(r)
-        sn1 = np.sin(asr * (1.0 - x) / 2.0)
-        sn2 = np.sin(asr * (1.0 + x) / 2.0)
-        dim1w = np.outer(dim1, w)
-        hkdim2 = np.outer(hk, dim2)
-        hsdim2 = np.outer(hs, dim2)
-        dim1sn1 = np.outer(dim1, sn1)
-        dim1sn2 = np.outer(dim1, sn2)
-        sn12 = np.multiply(sn1, sn1)
-        sn22 = np.multiply(sn2, sn2)
-        bvn = asr * np.sum(np.multiply(dim1w, np.exp(np.divide(np.multiply(dim1sn1, hkdim2) - hsdim2,
-                                                               (1 - np.outer(dim1, sn12))))) +
-                           np.multiply(dim1w, np.exp(np.divide(np.multiply(dim1sn2, hkdim2) - hsdim2,
-                                                               (1 - np.outer(dim1, sn22))))), axis=1) / (4 * np.pi) \
-              + np.multiply(_norm_cdf(-dh), _norm_cdf(-dk))
-    else:
-        if r < 0:
-            dk = -dk
-            hk = -hk
-
-        if abs(r) < 1:
-            opmr = (1.0 - r) * (1.0 + r)
-            sopmr = np.sqrt(opmr)
-            xmy2 = np.multiply(dh - dk, dh - dk)
-            xmy = np.sqrt(xmy2)
-            rhk8 = (4.0 - hk) / 8.0
-            rhk16 = (12.0 - hk) / 16.0
-            asr = -1.0 * (np.divide(xmy2, opmr) + hk) / 2.0
-
-            ind = asr > 100
-            bvn[ind] = sopmr * np.multiply(np.exp(asr[ind]),
-                                           1.0 - np.multiply(np.multiply(rhk8[ind], xmy2[ind] - opmr),
-                                                             (1.0 - np.multiply(rhk16[ind], xmy2[ind]) / 5.0) / 3.0)
-                                           + np.multiply(rhk8[ind], rhk16[ind]) * opmr * opmr / 5.0)
-
-            ind = hk > -100
-            ncdfxmyt = np.sqrt(2.0 * np.pi) * _norm_cdf(-xmy / sopmr)
-            bvn[ind] = bvn[ind] - np.multiply(np.multiply(np.multiply(np.exp(-hk[ind] / 2.0), ncdfxmyt[ind]), xmy[ind]),
-                                              1.0 - np.multiply(np.multiply(rhk8[ind], xmy2[ind]),
-                                                                (1.0 - np.multiply(rhk16[ind], xmy2[ind]) / 5.0) / 3.0))
-            sopmr = sopmr / 2
-            for ix in [-1, 1]:
-                xs = np.multiply(sopmr + sopmr * ix * x, sopmr + sopmr * ix * x)
-                rs = np.sqrt(1 - xs)
-                xmy2dim2 = np.outer(xmy2, dim2)
-                dim1xs = np.outer(dim1, xs)
-                dim1rs = np.outer(dim1, rs)
-                dim1w = np.outer(dim1, w)
-                rhk16dim2 = np.outer(rhk16, dim2)
-                hkdim2 = np.outer(hk, dim2)
-                asr1 = -1.0 * (np.divide(xmy2dim2, dim1xs) + hkdim2) / 2.0
-
-                ind1 = asr1 > -100
-                cdim2 = np.outer(rhk8, dim2)
-                sp1 = 1.0 + np.multiply(np.multiply(cdim2, dim1xs), 1.0 + np.multiply(rhk16dim2, dim1xs))
-                ep1 = np.divide(np.exp(np.divide(-np.multiply(hkdim2, (1.0 - dim1rs)),
-                                                 2.0 * (1.0 + dim1rs))), dim1rs)
-                bvn = bvn + np.sum(np.multiply(np.multiply(np.multiply(sopmr, dim1w), np.exp(np.multiply(asr1, ind1))),
-                                               np.multiply(ep1, ind1) - np.multiply(sp1, ind1)), axis=1)
-            bvn = -bvn / (2.0 * np.pi)
-
-        if r > 0:
-            bvn = bvn + _norm_cdf(-np.maximum(dh, dk))
-        elif r < 0:
-            bvn = -bvn + np.maximum(0, _norm_cdf(-dh) - _norm_cdf(-dk))
-
-    return bvn
-
-
-def _gauss_legendre_quad(r):
-    """
-    Return weights and abscissae for the Legendre-Gauss quadrature integral approximation
-    :param r: correlation
-    :return:
-    """
-    if np.abs(r) < 0.3:
-        lg = 3
-        w = np.array([0.1713244923791705, 0.3607615730481384, 0.4679139345726904])
-        x = np.array([0.9324695142031522, 0.6612093864662647, 0.2386191860831970])
-    elif np.abs(r) < 0.75:
-        lg = 6
-        w = np.array([.04717533638651177, 0.1069393259953183, 0.1600783285433464,
-                      0.2031674267230659, 0.2334925365383547, 0.2491470458134029])
-        x = np.array([0.9815606342467191, 0.9041172563704750, 0.7699026741943050,
-                      0.5873179542866171, 0.3678314989981802, 0.1252334085114692])
-    else:
-        lg = 10
-        w = np.array([0.01761400713915212, 0.04060142980038694, 0.06267204833410906,
-                      0.08327674157670475, 0.1019301198172404, 0.1181945319615184,
-                      0.1316886384491766, 0.1420961093183821, 0.1491729864726037,
-                      0.1527533871307259])
-        x = np.array([0.9931285991850949, 0.9639719272779138, 0.9122344282513259,
-                      0.8391169718222188, 0.7463319064601508, 0.6360536807265150,
-                      0.5108670019508271, 0.3737060887154196, 0.2277858511416451,
-                      0.07652652113349733])
-
-    return lg, w, x
-
-
-"""
-Weight functions:
-
-A valid weight is a python function of the form weight(birth, persistence, **kwargs) defining a scalar-valued function over the birth-persistence plane, where birth and persistence are numpy arrays of equal length. To ensure stability, functions should vanish continuously at the
-line persistence = 0 (see [PI] for details).
-"""
-
-def linear_ramp(birth, pers, low=0.0, high=1.0, start=0.0, end=1.0):
-    """
-    Continuous peicewise linear ramp function which is constant below and above specified input values
-    :param birth: birth coordinates
-    :param pers: persistence coordinates
-    :param low: minimal weight
-    :param high: maximal weight
-    :param start: start persistence value of linear transition from low to high weight
-    :param end: end persistence value of linear transition from low to high weight
-    :return: weight at persistence pair
-    """
-    try:
-        n = len(birth)
-    except:
-        n = 1
-        birth = [birth]
-        pers = [pers]
-
-    w = np.zeros((n,))
-    for i in range(n):
-        if pers[i] < start:
-            w[i] = low
-        elif pers[i] > end:
-            w[i] = high
-        else:
-            w[i] = (pers[i] - start) * (high - low) / (end - start) + low
-
-    return w
-
-def persistence(birth, pers, n=1.0):
-    """
-    Continuous monotonic function which weight a persistence pair (b,p) by p^n for some n > 0
-    :param birth: birth coordinates
-    :param pers: persistence coordinates
-    :param n: positive float
-    :return: weight at persistence pair
-    """
-    return pers ** n
